requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==1.11.6
django-filter==2.1.0
djangorestframework==3.9.1
idna==2.8
Markdown==3.0.1
Pillow==5.4.0
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
djangorestframework-jwt==1.11.0
